// src/app/api/graphql-types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
  I18NLocaleCode: { input: any; output: any; }
  JSON: { input: any; output: any; }
  PageContentDynamicZoneInput: { input: any; output: any; }
  Upload: { input: any; output: any; }
};
// ...
export type ComponentLayoutImageGrid = {
  __typename?: 'ComponentLayoutImageGrid';
  id: Scalars['ID']['output'];
  images?: Maybe<Array<Maybe<ComponentMediaImage>>>;
  style?: Maybe<Enum_Componentlayoutimagegrid_Style>;
};
// ...
export type ComponentMediaImage = {
  __typename?: 'ComponentMediaImage';
  alt_text: Scalars['String']['output'];
  description?: Maybe<Scalars['String']['output']>;
  file: UploadFileEntityResponse;
  id: Scalars['ID']['output'];
};
// ...
export enum Enum_Componentlayoutimagegrid_Style {
  Auto = 'auto',
  LandscapeXLandscape = 'landscape_x_landscape',
  LandscapeXPortrait = 'landscape_x_portrait',
  PortraitXLandscape = 'portrait_x_landscape',
  PortraitXPortrait = 'portrait_x_portrait'
}
// ...
export type UploadFile = {
  __typename?: 'UploadFile';
  alternativeText?: Maybe<Scalars['String']['output']>;
  caption?: Maybe<Scalars['String']['output']>;
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  ext?: Maybe<Scalars['String']['output']>;
  formats?: Maybe<Scalars['JSON']['output']>;
  hash: Scalars['String']['output'];
  height?: Maybe<Scalars['Int']['output']>;
  mime: Scalars['String']['output'];
  name: Scalars['String']['output'];
  previewUrl?: Maybe<Scalars['String']['output']>;
  provider: Scalars['String']['output'];
  provider_metadata?: Maybe<Scalars['JSON']['output']>;
  related?: Maybe<Array<Maybe<GenericMorph>>>;
  size: Scalars['Float']['output'];
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
  url: Scalars['String']['output'];
  width?: Maybe<Scalars['Int']['output']>;
};

export type UploadFileEntity = {
  __typename?: 'UploadFileEntity';
  attributes?: Maybe<UploadFile>;
  id?: Maybe<Scalars['ID']['output']>;
};

export type UploadFileEntityResponse = {
  __typename?: 'UploadFileEntityResponse';
  data?: Maybe<UploadFileEntity>;
};